+++ src/pmp/visualization/MeshViewer.h
@@ -4,7 +4,8 @@
 #pragma once
 
 #include "pmp/visualization/TrackballViewer.h"
-#include "pmp/visualization/SurfaceMeshGL.h"
+#include "pmp/SurfaceMesh.h"
+#include "pmp/visualization/Renderer.h"
 
 namespace pmp {
 
@@ -53,7 +54,7 @@ public:
 
 protected:
     SurfaceMesh mesh_;
-    SurfaceMeshGL renderer_;
+    Renderer renderer_;
     std::string filename_; //!< the current file
     float crease_angle_;
 };