+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 
 ### Changed
 
+- `SurfaceMeshGL` renamed to `Renderer` and de-coupled from `SurfaceMesh`
 - Remove max angle triangulation objective leading to fold-overs.
 - Breaking change: Re-design algorithms API to use a simple function interface.
 - Breaking change: Make helper classes `Heap`, `Quadric`, `NormalCone`, `TriangleKdTree` private.